model-training/requirements.txt
# Core ML and Data Science Libraries
tensorflow-macos==2.15.0
tensorflow-metal==1.1.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Image Processing
opencv-python-headless>=4.5.0
Pillow>=8.3.0
albumentations>=1.3.0

# Data Handling
kaggle>=1.5.12
wget>=3.2

# Progress and Visualization
tqdm>=4.62.0
plotly>=5.0.0

# Model Utilities
tensorboard>=2.8.0
keras-tuner>=1.1.0

# Jupyter (optional, for notebook development)
jupyter>=1.0.0
ipykernel>=6.0.0
